+++ src/main/java/insane96mcp/enhancedai/modules/illager/shoot/PillagerShoot.java
@@ -17,6 +17,7 @@ import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.ai.goal.RangedCrossbowAttackGoal;
 import net.minecraft.world.entity.monster.Pillager;
 import net.minecraftforge.event.entity.EntityJoinLevelEvent;
+import net.minecraftforge.eventbus.api.EventPriority;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 
 @Label(name = "Pillager Shoot", description = "Use the enhancedai:better_pillager_shoot entity type tag to add more skeletons that are affected by this feature")
@@ -43,7 +44,7 @@ public class PillagerShoot extends Feature {
 		super(module, enabledByDefault, canBeDisabled);
 	}
 
-	@SubscribeEvent
+	@SubscribeEvent(priority = EventPriority.LOWEST)
 	public void onSpawn(EntityJoinLevelEvent event) {
 		if (!this.isEnabled()
 				|| !(event.getEntity() instanceof Pillager pillager)